backend/requirements.txt
fastapi
uvicorn
langchain
chromadb
sentence-transformers
pydantic
python-dotenv
langchain-community
langchain-google-genai
